Rename user state to index
