+++ src/components/home/UserDisplay.tsx
@@ -5,7 +5,7 @@ import { Dialog, Transition } from '@headlessui/react';
 import Button from '../Button';
 import Input from '../Input';
 import SpaceBetween from '../../layouts/SpaceBetween';
-import { UserContext } from '../../states/user/user';
+import { UserContext } from '../../states/user';
 import useRenameUser from '../../states/user/renameUser';
 import ErrorCard from '../ErrorCard';
 import { isLenZero, notNil } from '../../utils';
